@@ -1,6 +1,6 @@
 {
   "name": "igo",
-  "version": "1.14.0",
+  "version": "1.14.1",
   "description": "Infrastructure géomatique ouverte",
   "author": "Communauté IGO",
   "repository": {
@@ -49,26 +49,26 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^14.1.2",
-    "@angular/cdk": "^14.1.1",
-    "@angular/common": "^14.1.2",
-    "@angular/compiler": "^14.1.2",
-    "@angular/core": "^14.1.2",
-    "@angular/forms": "^14.1.2",
-    "@angular/material": "^14.1.1",
-    "@angular/platform-browser": "^14.1.2",
-    "@angular/platform-browser-dynamic": "^14.1.2",
-    "@angular/router": "^14.1.2",
-    "@angular/service-worker": "^14.1.2",
+    "@angular/animations": "^14.2.12",
+    "@angular/cdk": "^14.2.7",
+    "@angular/common": "^14.2.12",
+    "@angular/compiler": "^14.2.12",
+    "@angular/core": "^14.2.12",
+    "@angular/forms": "^14.2.12",
+    "@angular/material": "^14.2.7",
+    "@angular/platform-browser": "^14.2.12",
+    "@angular/platform-browser-dynamic": "^14.2.12",
+    "@angular/router": "^14.2.12",
+    "@angular/service-worker": "^14.2.12",
     "@azure/msal-angular": "^2.4.1",
     "@azure/msal-browser": "^2.28.1",
-    "@igo2/auth": "1.14.0",
-    "@igo2/common": "1.14.0",
-    "@igo2/context": "1.14.0",
-    "@igo2/core": "1.14.0",
-    "@igo2/geo": "1.14.0",
-    "@igo2/integration": "1.14.0",
-    "@igo2/utils": "1.14.0",
+    "@igo2/auth": "1.14.1",
+    "@igo2/common": "1.14.1",
+    "@igo2/context": "1.14.1",
+    "@igo2/core": "1.14.1",
+    "@igo2/geo": "1.14.1",
+    "@igo2/integration": "1.14.1",
+    "@igo2/utils": "1.14.1",
     "@mat-datetimepicker/core": "~10.0.0",
     "@turf/buffer": "^6.5.0",
     "@turf/point-on-feature": "^6.5.0",
@@ -88,7 +88,7 @@
     "ngx-color-picker": "^13.0.0",
     "ngx-indexed-db": "^11.0.2",
     "ngx-toastr": "^15.0.0",
-    "ol": "^7.1.0",
+    "ol": "^7.2.2",
     "proj4": "^2.8.0",
     "rxjs": "^7.5.6",
     "scroll-into-view-if-needed": "^2.2.29",
@@ -101,14 +101,14 @@
     "zone.js": "~0.11.7"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^14.1.2",
-    "@angular-eslint/builder": "14.0.2",
-    "@angular-eslint/eslint-plugin": "14.0.2",
-    "@angular-eslint/eslint-plugin-template": "14.0.2",
-    "@angular-eslint/schematics": "14.0.2",
-    "@angular-eslint/template-parser": "14.0.2",
-    "@angular/cli": "^14.1.2",
-    "@angular/compiler-cli": "^14.1.2",
+    "@angular-devkit/build-angular": "^14.2.10",
+    "@angular-eslint/builder": "14.1.2",
+    "@angular-eslint/eslint-plugin": "14.1.2",
+    "@angular-eslint/eslint-plugin-template": "14.1.2",
+    "@angular-eslint/schematics": "14.1.2",
+    "@angular-eslint/template-parser": "14.1.2",
+    "@angular/cli": "^14.2.10",
+    "@angular/compiler-cli": "^14.2.10",
     "@babel/core": "^7.18.10",
     "@babel/preset-env": "^7.18.10",
     "@compodoc/compodoc": "^1.1.19",
@@ -125,10 +125,10 @@
     "eslint-plugin-unused-imports": "^2.0.0",
     "gulp": "^4.0.0",
     "gulp-babel": "^8.0.0",
+    "gulp-bump": "^3.2.0",
     "gulp-concat": "^2.6.1",
     "gulp-replace": "^1.1.3",
     "http-server": "^13.1.0",
-    "gulp-bump": "^3.2.0",
     "jasmine-core": "~3.8.0",
     "jasmine-spec-reporter": "~5.0.0",
     "karma": "^6.4.0",
