@@ -1,4 +1,9 @@
 import { environment as prodEnv } from './environment.github';
 
-prodEnv.igo.app.promotePwa = true;
+
+if (!prodEnv.igo.app.pwa) {
+    prodEnv.igo.app.pwa = {};
+}
+prodEnv.igo.app.pwa.enabled = true;
+prodEnv.igo.app.pwa.promote = true;
 export const environment = prodEnv;
